# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "MeshCore Analytics API",
    "version": "2.0.0",
    "description": "Runtime-validated contracts for privacy-filtered public, test, owner, and local operator endpoints."
  },
  "servers": [
    {
      "url": "https://ukmesh.com"
    }
  ],
...
  "paths": {
...
    "/api/telemetry/frontend-error": {
      "post": {
        "summary": "Create or run telemetry frontend error",
        "operationId": "post-telemetry-frontend-error",
        "tags": [
          "public"
        ],
        "x-access": "public",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid or out-of-bounds request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "code",
                    "requestId"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ownerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "meshcore_owner_session"
      },
      "operatorSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "meshcore_operator_session"
      },
      "operatorToken": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}